# docs/api-docs.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/health":{"get":{"operationId":"HealthController_health","parameters":[],"responses":{"200":{"description":""}},"tags":["Health"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/forgot-password":{"post":{"operationId":"AuthController_forgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/reset-password":{"post":{"operationId":"AuthController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/reset-password":{"get":{"operationId":"ResetPasswordPageController_serveResetPage","parameters":[{"name":"token","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["ResetPasswordPage"]}},"/users/me":{"get":{"operationId":"UsersController_getMe","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]},"put":{"operationId":"UsersController_updateMe","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]},"delete":{"operationId":"UsersController_deleteMe","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/me/password":{"put":{"operationId":"UsersController_changePassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/me/avatar":{"post":{"operationId":"UsersController_uploadAvatar","parameters":[],"responses":{"201":{"description":""}},"tags":["Users"]},"delete":{"operationId":"UsersController_deleteAvatar","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/destinations":{"get":{"operationId":"DestinationsController_findAll","parameters":[{"name":"q","required":true,"in":"query","schema":{"type":"string"}},{"name":"parentId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Destinations"]}},"/destinations/{id}":{"get":{"operationId":"DestinationsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Destinations"]}},"/restaurants":{"get":{"operationId":"RestaurantsController_findAll","parameters":[{"name":"destinationId","required":true,"in":"query","schema":{"type":"string"}},{"name":"parentDestinationId","required":true,"in":"query","schema":{"type":"string"}},{"name":"type","required":true,"in":"query","schema":{"type":"string"}},{"name":"kashrut","required":true,"in":"query","schema":{"type":"string"}},{"name":"q","required":true,"in":"query","schema":{"type":"string"}},{"name":"lat","required":true,"in":"query","schema":{"type":"string"}},{"name":"lng","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Restaurants"]}},"/restaurants/search":{"get":{"operationId":"RestaurantsController_aiSearch","parameters":[{"name":"destinationId","required":true,"in":"query","schema":{"type":"number"}},{"name":"q","required":true,"in":"query","schema":{"type":"string"}},{"name":"lat","required":true,"in":"query","schema":{"type":"string"}},{"name":"lng","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Restaurants"]}},"/restaurants/search/feedback":{"post":{"operationId":"RestaurantsController_recordFeedback","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecordFeedbackDto"}}}},"responses":{"201":{"description":""}},"tags":["Restaurants"]}},"/restaurants/nearby":{"get":{"operationId":"RestaurantsController_findNearby","parameters":[{"name":"lat","required":true,"in":"query","schema":{"type":"string"}},{"name":"lng","required":true,"in":"query","schema":{"type":"string"}},{"name":"limit","required":true,"in":"query","schema":{"type":"string"}},{"name":"kashrut","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Restaurants"]}},"/restaurants/{id}":{"get":{"operationId":"RestaurantsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Restaurants"]}},"/restaurants/import-json":{"post":{"operationId":"RestaurantsController_importFromJson","parameters":[],"responses":{"201":{"description":""}},"tags":["Restaurants"]}},"/restaurants/import-google":{"post":{"operationId":"RestaurantsController_importFromGoogle","parameters":[],"responses":{"201":{"description":""}},"tags":["Restaurants"]}},"/restaurants/reclassify":{"post":{"operationId":"RestaurantsController_reclassifyExisting","parameters":[],"responses":{"201":{"description":""}},"tags":["Restaurants"]}},"/restaurants/import-batch":{"post":{"operationId":"RestaurantsController_importBatch","parameters":[],"responses":{"201":{"description":""}},"tags":["Restaurants"]}},"/synagogues":{"get":{"operationId":"SynagoguesController_findByDestination","parameters":[{"name":"destinationId","required":true,"in":"query","schema":{"type":"string"}},{"name":"denomination","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Synagogues"]}},"/synagogues/{id}":{"get":{"operationId":"SynagoguesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Synagogues"]}},"/admin/destinations":{"post":{"operationId":"AdminController_createDestination","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDestinationDto"}}}},"responses":{"201":{"description":""}},"tags":["Admin"]}},"/admin/destinations/{id}":{"delete":{"operationId":"AdminController_deleteDestination","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Admin"]}},"/admin/restaurants":{"post":{"operationId":"AdminController_createRestaurant","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRestaurantDto"}}}},"responses":{"201":{"description":""}},"tags":["Admin"]}},"/admin/restaurants/{id}":{"delete":{"operationId":"AdminController_deleteRestaurant","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Admin"]}},"/admin/destinations/{id}/sync":{"post":{"operationId":"AdminController_syncDestination","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Admin"]}},"/admin/users/{id}":{"delete":{"operationId":"AdminController_blockUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Admin"]}},"/admin/messages/{id}":{"delete":{"operationId":"AdminController_deleteMessage","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Admin"]}},"/admin/candidates":{"get":{"operationId":"AdminController_listCandidateSynagogues","parameters":[{"name":"destinationId","required":true,"in":"query","schema":{"type":"number"}},{"name":"status","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Admin"]}},"/admin/candidates/{id}/approve":{"post":{"operationId":"AdminController_approveCandidateSynagogue","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Admin"]}},"/admin/candidates/{id}/reject":{"post":{"operationId":"AdminController_rejectCandidateSynagogue","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Admin"]}},"/admin/synagogues/{id}":{"delete":{"operationId":"AdminController_deleteSynagogue","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Admin"]}},"/admin/synagogues/regeocodeDestination/{destinationId}":{"post":{"operationId":"AdminController_regeocodeDestination","parameters":[{"name":"destinationId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Admin"]}},"/admin/synagogues/bulk":{"post":{"operationId":"AdminController_bulkImportSynagogues","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"responses":{"200":{"description":""}},"tags":["Admin"]}},"/minyans/mine":{"get":{"operationId":"MinyansController_findMine","parameters":[],"responses":{"200":{"description":""}},"tags":["Minyans"]}},"/minyans":{"get":{"operationId":"MinyansController_findAll","parameters":[{"name":"destinationId","required":true,"in":"query","schema":{"type":"number"}},{"name":"prayerType","required":true,"in":"query","schema":{"type":"string"}},{"name":"date","required":true,"in":"query","schema":{"type":"string"}},{"name":"lat","required":true,"in":"query","schema":{"type":"string"}},{"name":"lng","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Minyans"]},"post":{"operationId":"MinyansController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMinyanDto"}}}},"responses":{"201":{"description":""}},"tags":["Minyans"]}},"/minyans/{id}":{"get":{"operationId":"MinyansController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Minyans"]},"delete":{"operationId":"MinyansController_deleteMinyan","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Minyans"]}},"/minyans/{id}/register":{"post":{"operationId":"MinyansController_register","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Minyans"]},"delete":{"operationId":"MinyansController_unregister","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Minyans"]}},"/hosting/offers":{"post":{"operationId":"HostingController_createOffer","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOfferDto"}}}},"responses":{"201":{"description":""}},"tags":["Hosting"]}},"/hosting/offers/mine":{"get":{"operationId":"HostingController_myOffers","parameters":[],"responses":{"200":{"description":""}},"tags":["Hosting"]}},"/hosting/offers/{id}/deactivate":{"post":{"operationId":"HostingController_deactivateOffer","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Hosting"]}},"/hosting/offers/search":{"get":{"operationId":"HostingController_searchOffers","parameters":[{"name":"destinationId","required":true,"in":"query","schema":{"type":"number"}},{"name":"arrivalDate","required":true,"in":"query","schema":{"type":"string"}},{"name":"departureDate","required":true,"in":"query","schema":{"type":"string"}},{"name":"guestsCount","required":true,"in":"query","schema":{"type":"string"}},{"name":"forShabbat","required":true,"in":"query","schema":{"type":"string"}},{"name":"withChildren","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Hosting"]}},"/hosting/requests":{"post":{"operationId":"HostingController_createRequest","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRequestDto"}}}},"responses":{"201":{"description":""}},"tags":["Hosting"]}},"/hosting/requests/mine":{"get":{"operationId":"HostingController_myRequests","parameters":[],"responses":{"200":{"description":""}},"tags":["Hosting"]}},"/hosting/requests/received":{"get":{"operationId":"HostingController_requestsReceived","parameters":[],"responses":{"200":{"description":""}},"tags":["Hosting"]}},"/hosting/requests/{id}/approve":{"post":{"operationId":"HostingController_approve","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Hosting"]}},"/hosting/requests/{id}/reject":{"post":{"operationId":"HostingController_reject","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Hosting"]}},"/search/classify":{"get":{"operationId":"SearchController_classifyText","parameters":[{"name":"text","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Search"]}},"/search":{"post":{"operationId":"SearchController_search","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchDto"}}}},"responses":{"201":{"description":""}},"tags":["Search"]}}},"info":{"title":"Jewish On The Way — API","description":"Backend API for kosher restaurants, synagogues, minyans and Shabbat hosting worldwide","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{}},"LoginDto":{"type":"object","properties":{}},"ForgotPasswordDto":{"type":"object","properties":{}},"ResetPasswordDto":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{}},"ChangePasswordDto":{"type":"object","properties":{}},"RecordFeedbackDto":{"type":"object","properties":{}},"CreateDestinationDto":{"type":"object","properties":{}},"CreateRestaurantDto":{"type":"object","properties":{}},"CreateMinyanDto":{"type":"object","properties":{}},"CreateOfferDto":{"type":"object","properties":{}},"CreateRequestDto":{"type":"object","properties":{}},"SearchDto":{"type":"object","properties":{}}}}}